+++ src/Avalanche.Domain/Settings/Expiration.cs
@@ -1,0 +1,33 @@
+﻿namespace Avalanche.Domain.Settings
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum Expiration
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        Never,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        OneMonth,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        ThreeMonths,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        SixMonths,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        TwelveMonths
+    }
+}