+++ src/Avalanche.Runner/TestCase.cs
@@ -1,4 +1,6 @@
-﻿namespace Avalanche.Runner
+﻿using System.Net;
+
+namespace Avalanche.Runner
 {
     public class TestCase : TestCaseConfig
     {
@@ -16,10 +18,61 @@
         /// The config that is run for initialization
         /// </summary>
         public InitConfig Init { get; set; }
+
+        public Authorization Authorization { get; set; }
     }
 
     public class InitConfig
     {
         public string Url { get; set; }
+    }
+
+    /// <summary>
+    /// Authorization node for the configuration
+    /// </summary>
+    public class Authorization
+    {
+        /// <summary>
+        /// Supported values:
+        /// - oauth
+        /// </summary>
+        public string Type { get; set; } = "oauth";
+
+        /// <summary>
+        /// Supported GrantTypes:
+        /// - password
+        /// - client_credentials
+        /// </summary>
+        public string GrantType { get; set; }
+
+        /// <summary>
+        /// URL of the Authority to authenticate agains
+        /// </summary>
+        public string Authority { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public string ClientId { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public string ClientSecret { get; set; }
+
+        /// <summary>
+        /// Username for password grant_type
+        /// </summary>
+        public string Username { get; set; }
+
+        /// <summary>
+        /// Password for password grant_type
+        /// </summary>
+        public string Password { get; set; }
+
+        /// <summary>
+        /// Scopes for the authentication
+        /// </summary>
+        public string Scope { get; set; }
     }
 }