+++ src/Avalanche/Controllers/ScenarioController.cs
@@ -5,8 +5,6 @@ using Avalanche.ReadModel.QueryHandlers;
 using Avalanche.WriteModel;
 using Broadcast;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using System.Xml.Linq;
 
 namespace Avalanche.Controllers
 {
@@ -28,7 +26,7 @@ namespace Avalanche.Controllers
             var path = PathMapper.GetScenarioFile(name);
 
             var tsr = new ScenarioReader(_loggerFactory);
-            var settings = tsr.GetScenario(path);
+            var settings = tsr.GetScenarioFromFile(path);
 
             var trh = new TestRunQueryHandler(_builder);
             var runs = trh.Get(new ReadModel.Queries.GetTestsQuery { Scenario = name });
@@ -52,7 +50,7 @@ namespace Avalanche.Controllers
             var path = PathMapper.GetScenarioFile(name);
 
             var tsr = new ScenarioReader(_loggerFactory);
-            var settings = tsr.GetScenario(path);
+            var settings = tsr.GetScenarioFromFile(path);
 
             var trh = new TestRunQueryHandler(_builder);
             var stats = trh.Get(new ReadModel.Queries.GetTestsStatisticsQuery {  Scenario = name });