+++ Changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## vNext
+## v0.0.4
+### Added
+- ApiKeys can be generated and used to start a testrun
+- Added OAuth Authentication for testruns
+- Added a basic documentation
+- The runner can read scenarios from the STDIN
+
 ### Changed
-- 
+- Display an error if the scanario can't be saved